new custom asset fields +migration
